requirements.txt
playwright
python-dotenv 
supabase
pillow
google-cloud-storage